pyproject.toml
[project]
name = "rm4pro"
version = "0.1.0"
description = ""
authors = [
    {name = "Moti Elmakyes",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk (>=1.7.0,<2.0.0)",
    "broadlink (>=0.19.0,<0.20.0)",
    "rich (>=14.0.0,<15.0.0)",
    "bleak (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
